Added Extension clause for tar files
pathinfo's PATHINFO_EXTENSION uses something similar to  substr(strrchr('file.tar.gz', '.'),1) to find the extension.  This works fine for most files, but if it is say a gzipped tar file, it returns gz as the extension.
I added a little check to see if the file is a tar file.  This way we don't get files named filetar.bz2 or filetar.7z.